+++ src/morphocut/torch.py
@@ -34,8 +34,6 @@ class PyTorch(Node):
         module (torch.nn.Module): PyTorch module.
         input (input, Variable): Input.
         device (str or torch.device, optional): Device.
-        n_parallel (int, optional): Run multiple computations in parallel.
-            0 means synchronous computations.
         is_batch (bool, optional): Assume that input is a batch.
         output_key (optional): If the module has multiple outputs, output_key selects one of them.
         pin_memory (bool, optional): Use pinned memory for faster CPU-GPU transfer.
@@ -56,17 +54,15 @@ class PyTorch(Node):
         self,
         module: "torch.nn.Module",
         input: RawOrVariable,
-        device: Union[None, str, torch.device] = None,
+        device: Union[None, str, "torch.device"] = None,
         is_batch=None,
         output_key=None,
         pin_memory=None,
         pre_transform: Optional[Callable] = None,
         post_transform: Optional[Callable] = None,
-        autocast=False,
+        autocast=None,
     ):
         super().__init__()
-
-        print("PyTorch.device: ", device)
 
         if device is not None:
             device = torch.device(device)